backend/requirements/local.txt
-r base.txt
pytest>=7.4.0
pytest-django>=4.7.0
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
